requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
pyotp==2.8.0
qrcode==7.4.2
Pillow==10.0.0
translatepy==2.3
requests==2.31.0
flask-socketio==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
bcrypt==4.0.1
cryptography==41.0.3